trying to find a way to make grid refinement on LBM and CUDA

+++ MR_LBM/src/var.h
@@ -12,7 +12,7 @@
 #include <cstring>
 
 /* ----------------------------- PROBLEM DEFINE ---------------------------- */
-typedef float dfloat;
+typedef double dfloat;
 
 #define GPU_INDEX 0
 /* --------------------------  SIMULATION DEFINES -------------------------- */
